calendar-app/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.5
requests==2.28.2
PyJWT==2.6.0
python-dotenv==1.0.0
gunicorn==20.1.0

